% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midlines_polish.R
\name{midlines_debit}
\alias{midlines_debit}
\title{Removes small groups of lines}
\usage{
midlines_debit(x, length)
}
\arguments{
\item{x}{a feature collection of sf linestrings. The input is intended to be the output of \code{\link{midlines_clean}}}

\item{length}{the length below which groups of lines will be removed}
}
\description{
Lines are removed if, when grouped into contiguous clusters, the total length of the cluster is less the specified length. The dataset of ungrouped lines is returned.
}
