% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midlines_polish.R
\name{midlines_dedensify}
\alias{midlines_dedensify}
\title{Removes points on lines between nodes}
\usage{
midlines_dedensify(x, density)
}
\arguments{
\item{x}{a feature collection of sf linestrings. The input is intended to be the output of \code{\link{midlines_clean}}}

\item{density}{is the desired distance between points. This is passed to \code{\link[sf]{st_line_sample}}}
}
\description{
Without modifying the nodes where lines meet, the lines between those nodes have their points reduced to de-densify. This might be useful to reduce the size of the line collection. The intention is that a high density of points can be used to estimate the midlines but this can be reduced if desired. The dataset of ungrouped lines is returned.
}
\examples{
library(sf)
poly = st_buffer(st_linestring(matrix(c(0,0,10,0,10,10,0,10,0,0),ncol=2, byrow=TRUE) ),0.75)
plot(poly, col = "GRAY")

ml = midlines_clean(midlines_draw(poly, dfMaxLength = 1))
ml = ml[ml$removed_flag==0,]
plot(ml$geometry, add = TRUE)

ml_dedensified = midlines_dedensify(ml, density = 1)
plot(poly, col = "GRAY")
plot(ml_dedensified$geometry, add = TRUE)
}
