% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midline_functions_2021-10-15.R
\name{midlines_draw}
\alias{midlines_draw}
\title{Estimates midlines of sf polygon(s)}
\usage{
midlines_draw(dat, boarder_line = NULL, max_dist)
}
\description{
Estimates midlines of sf polygon(s)
}
