% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midlines_polish.R
\name{midlines_smooth}
\alias{midlines_smooth}
\title{Uses a rolling mean to smooth midlines}
\usage{
midlines_smooth(x, width = 3)
}
\arguments{
\item{x}{Feature collection of sf linestrings. The input is intended to be the output of \code{\link{midlines_clean}}}

\item{width}{option passed to \code{zoo{rollapply}} "specifying the window width (in numbers of observations)" used to calculate the rolling mean}
}
\description{
Uses \code{zoo{rollapply}} to smooth lines, while keeping nodes where lines meet unchanged. The dataset of ungrouped lines is returned.
}
\examples{
library(sf)
poly = st_buffer(st_linestring(matrix(c(0,0,10,0,10,10,0,10,0,0),ncol=2, byrow=TRUE) ),0.75)
plot(poly, col = "GRAY")

ml = midlines_clean(midlines_draw(poly, dfMaxLength = 1))
ml = ml[ml$removed_flag==0,]
plot(ml$geometry, add = TRUE)

ml_smooth = midlines_smooth(ml)
plot(poly, col = "GRAY")
plot(ml_smooth$geometry, add = TRUE)
}
