% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thames}
\alias{thames}
\title{An sf collection containing a polygon of a section of the river Thames in London}
\format{
\subsection{\code{thames}}{

A data frame with 1 feature and 0 fields:
\describe{
\item{geometry}{the geometry column containing coordinates}
}
}
}
\source{
\url{https://www.openstreetmap.org}
}
\usage{
thames
}
\description{
Data adapted from that available from OpenStreetMap
}
\keyword{datasets}
