% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midlines_refine.R
\name{midlines_group}
\alias{midlines_group}
\title{Groups lines into multilinestring of clusters of lines which are all contiguous.}
\usage{
midlines_group(x)
}
\arguments{
\item{x}{an sf linestring or collection of linestrings}
}
\description{
Groups lines into multilinestring of clusters of lines which are all contiguous.
}
